requirements.txt
Pillow>=10.0.0
tqdm>=4.66.0
Flask>=2.3.0